snapped off per-sample SV bits
